Revert "Anadida barra al componente del mapa"
This reverts commit b0b1d6079ee032d4a8e8946388e15279ee019de6.
